+++ frontend/src/App.tsx
@@ -1,10 +1,14 @@
 import { Route, Routes } from "react-router-dom";
+
+import { Layout } from "./components/layout/Layout";
 import Home from "./routes/Home";
 
 export default function App() {
   return (
     <Routes>
-      <Route path="/" element={<Home />} />
+      <Route element={<Layout />}>
+        <Route path="/" element={<Home />} />
+      </Route>
     </Routes>
   );
 }